+++ common/src/main/java/software/bernie/geckolib/renderer/GeoObjectRenderer.java
@@ -5,7 +5,6 @@ import com.mojang.blaze3d.vertex.VertexConsumer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
-import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
 import org.joml.Matrix4f;
 import software.bernie.geckolib.GeckoLibServices;
@@ -96,7 +95,6 @@ public class GeoObjectRenderer<T extends GeoAnimatable> implements GeoRenderer<T
 	 * @param buffer The VertexConsumer to use for rendering, or null to use the default for the RenderType
 	 * @param packedLight The light level at the given render position for rendering
 	 */
-	@ApiStatus.Internal
 	public void render(PoseStack poseStack, T animatable, @Nullable MultiBufferSource bufferSource, @Nullable RenderType renderType,
 					   @Nullable VertexConsumer buffer, int packedLight, float partialTick) {
 		if (bufferSource == null)