+++ fabric/src/main/java/software/bernie/geckolib/platform/GeckoLibClientFabric.java
@@ -68,6 +68,6 @@ public class GeckoLibClientFabric implements GeckoLibClient {
      */
     @Override
     public int getDyedItemColor(ItemStack itemStack, int defaultColor) {
-        return itemStack.is(ItemTags.DYEABLE) ? DyedItemColor.getOrDefault(itemStack, 0) : 0;
+        return itemStack.is(ItemTags.DYEABLE) ? DyedItemColor.getOrDefault(itemStack, defaultColor) : defaultColor;
     }
 }